fix cp

+++ init.sh
@@ -1,12 +1,31 @@
 DIR=$(basename $(dirname $(pwd)))
 
-cp .clasp.template.json ../
-cp .gitignore ../
-cp package.json ../
-cp yarn.lock ../
-cp tslint.json ../
-cp -r src ../
+if [! -f '.clasp.template.json'] && [! -f'.clasp.json']; then
+  cp .clasp.template.json ../
+fi
 
-cd ../
+if [! -f '.gitignore']; then
+  cp .gitignore ../
+fi
 
-git remote set-url origin git@github.com:kobakazu0429/${DIR}.git
+if [! -f 'package.json']; then
+  cp package.json ../
+fi
+
+if [! -f 'yarn.lock']; then
+  cp yarn.lock ../
+fi
+
+if [! -f 'tslint.json']; then
+  cp tslint.json ../
+fi
+
+if [! -d 'src/']; then
+  cp -r src ../
+fi
+
+
+if [ -d '.git/' ]; then
+  cd ../
+  git remote set-url origin git@github.com:kobakazu0429/${DIR}.git
+fi
